requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.1
requests==2.31.0
db-dtypes==1.2.0
alpaca-py==0.16.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.19.0